Override default task to [:cucumber, :rails]
It seems that using minitest-rails with cucumber causes problems to the
latter, as cucumber could not clear the test database as another user is
connected to it.

Quickest fix was to redefine the default task to simply call :cucumber
and then :test.  This isolates both and prevents the error encountered,
though this may mean that the test database is cleared twice.